refactor(backend-jwt): move modules for architecture-consistency
